improve for basic commands
