+++ iina/zh-Hans.lproj/KeyBinding.strings
@@ -7,9 +7,22 @@
 */
 
 "cmd.ignore" = "忽略";
-
 "cmd.seek" = "查找";
-
 "cmd.set" = "设置属性";
-
 "cmd.cycle" = "循环属性";
+"cmd.add" = "增减属性";
+"cmd.multiply" = "循环属性";
+
+"to" = "为";
+"by" = "倍数";
+"add" = "增加";
+"minus" = "减少";
+
+"opt.pause" = "暂停";
+"opt.volume" = "音量";
+"opt.mute" = "静音";
+"opt.speed" = "速度";
+"opt.fullscreen" = "全屏";
+"opt.sub-delay" = "字幕延迟";
+"opt.sub-pos" = "字幕位置";
+"opt.sub-scale" = "字幕缩放";
